Modification in Submit Query API
