+++ src/main/java/com/naval_innovators/your_exam_sathi/YourExamSathiApplication.java
@@ -10,5 +10,4 @@ public class YourExamSathiApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(YourExamSathiApplication.class, args);
 	}
-
 }
