--- src/main/java/com/naval_innovators/your_exam_sathi/auth_service/controller/CourseController.java
@@ -1,31 +1,0 @@
-package com.naval_innovators.your_exam_sathi.auth_service.controller;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.naval_innovators.your_exam_sathi.auth_service.models.Course;
-import com.naval_innovators.your_exam_sathi.auth_service.service.CourseService;
-
-import lombok.RequiredArgsConstructor;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import java.util.*;
-
-
-@RestController
-@RequestMapping("/api/courses")
-@RequiredArgsConstructor
-public class CourseController {
-    private final CourseService courseService;
-
-    @GetMapping("/all")
-    public ResponseEntity<List<Course>> getAllCourses() {
-        List<Course> courses = courseService.findAllCourses();
-        if(courses.isEmpty()){
-            return ResponseEntity.noContent().build(); // response code - 204(No Content)
-        }
-        return ResponseEntity.ok(courses);
-    }
-    
-}
