--- src/main/java/com/naval_innovators/your_exam_sathi/auth_service/repository/SubjectRepository.java
@@ -1,9 +1,0 @@
-package com.naval_innovators.your_exam_sathi.auth_service.repository;
-
-import com.naval_innovators.your_exam_sathi.auth_service.models.Subject;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface SubjectRepository extends JpaRepository<Subject, Long> {
-}
