--- src/main/java/com/naval_innovators/your_exam_sathi/auth_service/service/CourseService.java
@@ -1,21 +1,0 @@
-package com.naval_innovators.your_exam_sathi.auth_service.service;
-
-import com.naval_innovators.your_exam_sathi.auth_service.models.Course;
-import com.naval_innovators.your_exam_sathi.auth_service.repository.CourseRepository;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-public class CourseService {
-
-    private final CourseRepository courseRepository;
-
-    public CourseService(CourseRepository courseRepository) {
-        this.courseRepository = courseRepository;
-    }
-
-    public List<Course> findAllCourses() {
-        return courseRepository.findAll();
-    }
-}
